+++ Julia_Dementieva/my-app-messenger/src/components/Messenger/Messenger.tsx
@@ -1,10 +1,10 @@
-import React from 'react'
+import React from 'react';
 import {ListItem, ListItemText, ListItemAvatar, Avatar, CircularProgress } from '@material-ui/core';
 import {MessageList} from '../MessageList';
 import {MessageForm} from '../MessageForm';
 import {Error} from '../../pages/Error';
 
-import {MessageType} from '../../types/types';
+import {MessageFullInfoType, MessageType} from '../../types/types';
 
 import './Messenger.css';
 
@@ -14,7 +14,7 @@ type MessengerType = {
     authorChat: string;
     namePerson: string;
     avatarChat: string;
-    messages: Array<MessageType>;
+    messages: Array<MessageFullInfoType>;
     isLoading: boolean;
 };
